app/api/requirements.txt
flask
pytest
psycopg2-binary
simplejson